fix: TokenType for balances use NATIVE_TOKEN and Transaction type uses NATIVE_COIN to define native chain token
